chore: move botocore utils to utils_botocore
